Move: Inventory Matcher to Common Package
