+++ server/src/main/java/umm3601/Bootstrap.java
@@ -13,10 +13,10 @@ import umm3601.Auth.AuthController;
 import umm3601.Auth.PermissionsService;
 import umm3601.Auth.RouteRegistrar;
 import umm3601.Common.ApiExceptionHandler;
+import umm3601.Common.InventoryMatcher;
 import umm3601.Family.FamilyChecklistService;
 import umm3601.Family.FamilyController;
 import umm3601.Family.FamilyPortalController;
-import umm3601.Family.InventoryMatcher;
 import umm3601.Family.InventoryReservationService;
 import umm3601.Inventory.BarcodeController;
 import umm3601.Inventory.InventoryController;
