+++ server/src/test/java/umm3601/Common/InventoryMatcherSpec.java
@@ -1,4 +1,4 @@
-package umm3601.Family;
+package umm3601.Common;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
