+++ server/src/test/java/umm3601/Family/FamilyControllerSpec.java
@@ -52,6 +52,7 @@ import io.javalin.http.NotFoundResponse;
 import io.javalin.json.JavalinJackson;
 import io.javalin.validation.BodyValidator;
 import umm3601.Auth.Role;
+import umm3601.Common.InventoryMatcher;
 import umm3601.Family.Family.AvailabilityOptions;
 import umm3601.Family.Family.StudentInfo;
 import umm3601.Users.Users;
